+++ wx_main.py
@@ -23,6 +23,7 @@ from apricot.player.playback import PlaybackMixin
 from apricot.player.mpv import MpvMixin
 from apricot.network.youtube import YoutubeMixin
 from apricot.system.registry import RegistryMixin
+from apricot.system.diagnostics import DiagnosticsMixin
 from apricot.updater.updater import AppUpdaterMixin
 from apricot.data.manager import DataManagerMixin
 from apricot.utils import UtilsMixin
@@ -55,7 +56,7 @@ from apricot.locales import TEXT
 
 from apricot.constants import *
 
-class MainFrame(CookiesUI, DownloadsUI, EqualizerUI, EventsUI, ListsUI, MenusUI, MiscUI, PlayerUI, SearchUI, ShortcutsUI, SystemUI, SettingsMixin, DownloaderMixin, MediaMixin, LibraryMixin, SearchMixin, VolumeMixin, CookiesMixin, DialogsMixin, PlaybackMixin, MpvMixin, YoutubeMixin, RegistryMixin, AppUpdaterMixin, DataManagerMixin, UtilsMixin, wx.Frame):
+class MainFrame(CookiesUI, DownloadsUI, EqualizerUI, EventsUI, ListsUI, MenusUI, MiscUI, PlayerUI, SearchUI, ShortcutsUI, SystemUI, SettingsMixin, DownloaderMixin, MediaMixin, LibraryMixin, SearchMixin, VolumeMixin, CookiesMixin, DialogsMixin, PlaybackMixin, MpvMixin, YoutubeMixin, RegistryMixin, DiagnosticsMixin, AppUpdaterMixin, DataManagerMixin, UtilsMixin, wx.Frame):
     def __init__(self, start_hidden_in_tray: bool = False) -> None:
         super().__init__(None, title=WINDOW_TITLE, size=(950, 680))
         APP_DIR.mkdir(parents=True, exist_ok=True)